+++ packages/react/src/components/gdpr-consent-dialog/settings.tsx
@@ -2,19 +2,18 @@ import { Accordion } from "../accordion";
 import { Switch } from "../switch";
 import { Text } from "../text";
 import { CONSENT_LEVELS_WITHOUT_NONE } from "./constants";
-import { Dispatch } from "./state";
 import * as styles from "./style.css";
-import { ConsentSettings } from "./types";
+import { ConsentSettings, ConsentSettingsLevel } from "./types";
 
 type Props = {
   consentSettings: ConsentSettings;
-  dispatch: Dispatch;
+  toggleConsent: (level: ConsentSettingsLevel) => void;
   disabled?: boolean;
 };
 
 export const Settings = ({
   consentSettings,
-  dispatch,
+  toggleConsent,
   disabled = false,
 }: Props) => {
   return (
@@ -28,9 +27,7 @@ export const Settings = ({
             blocked={level === "necessary"}
             disabled={disabled}
             checked={consentSettings[level]}
-            onCheckedChange={() =>
-              dispatch({ type: "TOGGLE_CONSENT", payload: level })
-            }
+            onCheckedChange={() => toggleConsent(level)}
           />
         ),
         trigger: (
