+++ packages/react/src/components/gdpr-consent-dialog/state.ts
@@ -1,9 +1,11 @@
 import { GDPRConsent } from "@slashid/slashid";
 import { CONSENT_LEVELS_WITHOUT_NONE } from "./constants";
-import { ConsentSettings, ConsentSettingsLevel } from "./types";
+import { ActionType, ConsentSettings, ConsentSettingsLevel } from "./types";
+import { fromEntries } from "../utils";
 
 export type State = {
   consentSettings: ConsentSettings;
+  activeAction: ActionType;
   open: boolean;
   isLoading: boolean;
   hasError: boolean;
@@ -12,30 +14,33 @@ export type State = {
 
 type Action =
   | { type: "SET_OPEN"; payload: boolean }
-  | { type: "START_LOADING" }
-  | { type: "STOP_LOADING" }
+  | { type: "START_ACTION"; payload: ActionType }
+  | { type: "COMPLETE_ACTION" }
   | { type: "SET_HAS_ERROR"; payload: boolean }
   | { type: "SET_IS_CUSTOMIZING"; payload: boolean }
-  | { type: "SET_CONSENT_SETTINGS"; payload: GDPRConsent[] }
+  | { type: "SYNC_CONSENT_SETTINGS"; payload: GDPRConsent[] }
   | { type: "TOGGLE_CONSENT"; payload: ConsentSettingsLevel };
 
 export type Dispatch = React.Dispatch<Action>;
 
-export const mapConsentsToSettings = (consents: GDPRConsent[]) => {
-  const settings = Object.fromEntries(
-    CONSENT_LEVELS_WITHOUT_NONE.map((level) => [
-      level,
-      consents.map(({ consent_level }) => consent_level).includes(level),
-    ])
+export const mapConsentsToSettings = (
+  consents: GDPRConsent[]
+): ConsentSettings =>
+  fromEntries(
+    CONSENT_LEVELS_WITHOUT_NONE.map<[ConsentSettingsLevel, boolean]>(
+      (level) => [
+        level,
+        consents.map(({ consent_level }) => consent_level).includes(level),
+      ]
+    )
   );
-  return settings as ConsentSettings;
-};
 
 export const createInitialState = (
   initialConsentSettings: ConsentSettings,
   initialOpen: boolean
 ): State => ({
   consentSettings: initialConsentSettings,
+  activeAction: null,
   open: initialOpen,
   isLoading: false,
   hasError: false,
@@ -49,13 +54,14 @@ export const reducer = (state: State, action: Action): State => {
         ...state,
         open: action.payload,
       };
-    case "START_LOADING":
+    case "START_ACTION":
       return {
         ...state,
+        activeAction: action.payload,
         hasError: false,
         isLoading: true,
       };
-    case "STOP_LOADING":
+    case "COMPLETE_ACTION":
       return {
         ...state,
         isLoading: false,
@@ -71,10 +77,11 @@ export const reducer = (state: State, action: Action): State => {
         hasError: false,
         isCustomizing: action.payload,
       };
-    case "SET_CONSENT_SETTINGS":
+    case "SYNC_CONSENT_SETTINGS":
       return {
         ...state,
         consentSettings: mapConsentsToSettings(action.payload),
+        open: false,
       };
     case "TOGGLE_CONSENT":
       return {
