+++ packages/react/src/components/utils.ts
@@ -74,3 +74,11 @@ export const documentConstraints = {
   maxResolution: 64,
   maxSize: 10,
 };
+
+export const fromEntries = <
+  T extends ReadonlyArray<readonly [PropertyKey, unknown]>,
+>(
+  entries: T
+): { [K in T[number] as K[0]]: K[1] } => {
+  return Object.fromEntries(entries) as { [K in T[number] as K[0]]: K[1] };
+};
